# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MaintenancePizza API",
    "version": "2.0.0",
...
  },
  "servers": [
    { "url": "http://localhost:8000/api", "description": "Local dev server" }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/stores/{store}/tickets/{ticket}/warranties": {
      "post": {
        "tags": ["Issue Workflow"], "summary": "Add a warranty (multipart, optional files)", "operationId": "createWarranty",
        "description": "expiry_date is required — it is the date the warranty coverage expires.",
        "parameters": [{ "$ref": "#/components/parameters/Store" }, { "$ref": "#/components/parameters/Ticket" }],
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "required": ["ticket_issue_ids", "body", "expiry_date"], "properties": { "ticket_issue_ids[]": { "type": "array", "items": { "type": "integer" } }, "body": { "type": "string", "description": "Warranty description / terms" }, "expiry_date": { "type": "string", "format": "date", "description": "Date the warranty expires (YYYY-MM-DD)" }, "files[]": { "type": "array", "items": { "type": "string", "format": "binary" } } } } } } },
        "responses": {
          "201": { "description": "Warranty created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WarrantyResource" } } } },
          "401": { "$ref": "#/components/responses/Unauthenticated" }, "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "Sanctum personal access token issued by the auth service." }
    },
    "parameters": {
      "Store":       { "name": "store",       "in": "path", "required": true, "schema": { "type": "string" },  "description": "Store number string, e.g. 03795-00001", "example": "03795-00001" },
      "Ticket":      { "name": "ticket",      "in": "path", "required": true, "schema": { "type": "integer" }, "description": "Ticket id" },
...
    },
    "responses": {
      "Unauthenticated": {
        "description": "Missing or invalid bearer token",
        "content": { "application/json": { "schema": { "type": "object", "properties": { "message": { "type": "string", "example": "Unauthenticated." } } } } }
      },
      "ValidationError": {
        "description": "Request failed validation",
        "content": { "application/json": { "schema": { "type": "object", "required": ["message", "errors"], "properties": { "message": { "type": "string" }, "errors": { "type": "object", "additionalProperties": { "type": "array", "items": { "type": "string" } } } } } } }
      }
    },
    "schemas": {
...
      "User": {
        "type": "object", "description": "The author of a record.",
        "properties": { "id": { "type": "integer" }, "name": { "type": "string" }, "email": { "type": "string", "format": "email" } }
      },
...
      "Attachment": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "path": { "type": "string", "description": "Relative path on the public storage disk" },
          "url":  { "type": "string", "format": "uri", "description": "Full public URL — use this in <img src> / download links" },
          "original_name": { "type": "string", "nullable": true },
          "mime_type":     { "type": "string", "nullable": true },
          "size":          { "type": "integer", "nullable": true, "description": "File size in bytes" },
          "created_by":    { "type": "integer", "nullable": true },
          "created_at":    { "type": "string", "format": "date-time" }
        }
      },
...
      "Note": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "type": { "type": "string", "nullable": true, "description": "FinalNoteType value for ticket closing notes; null for generic notes on other entities." },
          "type_label": { "type": "string", "nullable": true, "description": "Human-readable label for `type` when it matches a FinalNoteType; null otherwise." },
          "body": { "type": "string" },
          "attachments": { "type": "array", "nullable": true, "items": { "$ref": "#/components/schemas/Attachment" }, "description": "Files attached to this note (e.g. a photo in a 'What we learned' note)." },
          "created_by": { "type": "integer", "nullable": true },
          "creator": { "allOf": [{ "$ref": "#/components/schemas/User" }], "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }, "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Warranty": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" }, "body": { "type": "string" },
          "expiry_date": { "type": "string", "format": "date", "description": "Date the warranty expires (YYYY-MM-DD)" },
          "attachments": { "type": "array", "nullable": true, "items": { "$ref": "#/components/schemas/Attachment" } },
          "notes":        { "type": "array", "nullable": true, "items": { "$ref": "#/components/schemas/Note" } },
          "ticket_issue_ids": { "type": "array", "nullable": true, "items": { "type": "integer" } },
          "created_by": { "type": "integer", "nullable": true }, "created_at": { "type": "string", "format": "date-time" }, "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "WarrantyResource": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Warranty" } } },
...
    }
  }
}